Assign modules -- updated

+++ src/app/api/master/rbac/route.ts
@@ -53,8 +53,21 @@ async function UpsertAssignModulesToRole(roleId: string, modulesWithPermissions:
 }
 
 async function upsertModulePermissions(roleId: string, moduleData: Module, assignments: any[]) {
+    // Check if the module has any submodules with non-zero permissions
+    const hasNonZeroChildPermission = moduleData.submodules && moduleData.submodules.some(submodule => submodule.permissions > 0);
+
+    // Check if the module was previously assigned (i.e., exists in the modulePermissions table for the role)
+    const existingModulePermission = await prisma.modulePermissions.findUnique({
+        where: {
+            roleId_moduleId: {
+                roleId: roleId,
+                moduleId: moduleData.moduleId,
+            },
+        },
+    });
+
     // Check if permissions are 0, in which case we remove the module access by deleting it
-    if (moduleData.permissions === 0) {
+    if (moduleData.permissions === 0  && !hasNonZeroChildPermission  && existingModulePermission) {
         // Delete the module permission for the role
         await prisma.modulePermissions.delete({
             where: {
@@ -89,8 +102,19 @@ async function upsertModulePermissions(roleId: string, moduleData: Module, assig
 }
 
 async function upsertSubmodulePermissions(roleId: string, moduleId: string, submodule: SubModules, assignments: any[]) {
-    if (submodule.permissions === 0) {
-        // If permission is 0, delete the submodule permission
+
+    const existingModulePermission = await prisma.modulePermissions.findUnique({
+        where: {
+            roleId_moduleId_submoduleId: {
+                roleId: roleId,
+                moduleId: moduleId,
+                submoduleId: submodule.submoduleId,
+            },
+        },
+    });
+
+    // If permission is 0 and already assigned, delete the submodule permission
+    if (submodule.permissions === 0 && existingModulePermission) {
         await prisma.modulePermissions.delete({
             where: {
                 roleId_moduleId_submoduleId: {
